+++ microsandbox-core/lib/management/home.rs
@@ -172,7 +172,7 @@ pub async fn install(
     image: &Reference,
     script: Option<&str>,
     alias: Option<&str>,
-    cpus: Option<u8>,
+    cpus: Option<f32>,
     memory: Option<u32>,
     volumes: Vec<String>,
     ports: Vec<String>,
@@ -217,7 +217,7 @@ pub async fn install(
         let mut b = Sandbox::builder().image(ReferenceOrPath::Reference(image.clone()));
 
         if let Some(cpus) = cpus {
-            b = b.cpus(cpus as f32);
+            b = b.cpus(cpus);
         }
 
         if let Some(memory) = memory {
